Add sad path for delete grain spec

+++ app/graphql/mutations/delete_grain.rb
@@ -4,9 +4,13 @@ module Mutations
     argument :id, ID, required: false
 
     def resolve(id:)
-      grain = Grain.find(id)
-      grain.destroy
-      { response: "You've successfully destroyed this grain." }
+      if Grain.exists?(id)
+        grain = Grain.find(id)
+        grain.destroy
+        { response: "You've successfully destroyed this grain." }
+      else
+        { response: "Grain not found" }
+      end
     end
   end
 end
